protectRoutes

+++ server/api/users.js
@@ -8,7 +8,7 @@ router.get('/', async(req, res, next) => { // api/users
     res.send(await User.findAll());
   }
   catch (ex) {
-    next (ex)
+    next(ex)
   }
 })
 
@@ -23,21 +23,26 @@ router.get('/get-user', (req, res, next) => {
 
 router.get('/:userId', async(req, res, next) => { // single user profile
   try {
-    const userProfile = await User.findOne({
-      where: {
-        id: req.params.userId
-      },
-      include: [Preference, Dog]
-    });
-    console.log('backend', userProfile)
-    res.send(userProfile)
+    if (req.user.id === Number(req.params.userId)){
+      const userProfile = await User.findOne({
+        where: {
+          id: req.params.userId
+        },
+        include: [Preference, Dog]
+      });
+      console.log('backend', userProfile)
+      res.send(userProfile)
+    }
+    else {
+      res.send('You are unauthorized to access this.')
+    }
   }
   catch (ex) {
     next(ex)
   }
 })
 
-router.post('/register', async(req,res,next) => { // register a user (api/users/register)
+router.post('/register', async(req, res, next) => { // register a user (api/users/register)
   try {
     console.log('trying to create a new user')
 
@@ -80,10 +85,15 @@ router.post('/register', async(req,res,next) => { // register a user (api/users/
   }
 })
 
-router.delete('/:userId', async(req,res,next) => { // delete a user (api/users)
+router.delete('/:userId', async(req, res, next) => { // delete a user (api/users)
   try {
-    await User.destroy({where: {id: req.params.userId}})
-    res.sendStatus(200)
+    if (req.user.id === Number(req.params.userId)){
+      await User.destroy({where: {id: req.params.userId}})
+      res.sendStatus(200)
+    }
+    else {
+      res.send('You are unauthorized to access this.')
+    }
   } catch (error) {
     console.log(error)
     res.sendStatus(500)
@@ -91,27 +101,32 @@ router.delete('/:userId', async(req,res,next) => { // delete a user (api/users)
 
 })
 
-router.put('/:userId', async(req,res,next) => { // update a user (api/users)
+router.put('/:userId', async(req, res, next) => { // update a user (api/users)
   try {
-    await Dog.update(req.body.dog, {
-      where: {
-        id: req.body.dog.id
-      }
-    })
-    const withoutDog = req.body;
-    delete withoutDog.dog
+    if (req.user.id === Number(req.params.userId)){
+      await Dog.update(req.body.dog, {
+        where: {
+          id: req.body.dog.id
+        }
+      })
+      const withoutDog = req.body;
+      delete withoutDog.dog
 
-    await User.update(withoutDog, {
-      where: {
-        id: req.params.userId
-      }
-    })
-    const updatedUser = await User.findOne({
-      where: {
-        id: req.params.userId
-      }
-    })
-    res.send(updatedUser);
+      await User.update(withoutDog, {
+        where: {
+          id: req.params.userId
+        }
+      })
+      const updatedUser = await User.findOne({
+        where: {
+          id: req.params.userId
+        }
+      })
+      res.send(updatedUser);
+    }
+  else {
+    res.send('You are unauthorized to access this.')
+  }
   } catch (error) {
     console.log(error)
     res.sendStatus(500)
